closetai/deployment/gpu-processor/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pillow>=10.0.0
numpy>=1.24.0
boto3>=1.28.0
python-multipart>=0.0.6
transformers>=4.30.0
accelerate>=0.20.0
torch>=2.0.0
